geoRouter better file structure
