+++ map/src/class/geoRouter/setters.js
@@ -1,3 +1,4 @@
+
 /*
     Saving profile params isn't easy process:
 
@@ -23,4 +24,12 @@ export function setParams({ opts }) {
     });
 
     this.flushState(next);
+}
+
+export function setPause({ pause }) {
+    // const next = this.nextState();
+    // next.paused = !!pause;
+    // this.flushState(next);
+
+this.flushState((o) => o.paused = !!pause);
 }
